+++ plugins/dsh-file-manage/scripts/bundle-client.mjs
@@ -2,7 +2,7 @@
 import { build } from 'esbuild'
 import { resolve } from 'node:path'
 
-const id = '@dsh-sparrow/dsh-file-session'
+const id = '@dsh-sparrow/dsh-file-manage'
 await build({
   entryPoints: [resolve(process.cwd(), 'src/client/index.ts')],
   outfile: resolve(process.cwd(), 'lib/client.js'),